fix(SGE): ~ST `Kardia` Targeting
UI and Hard Targets now both must be able to receive the action, and must be within range.
The range check may be a matter of opinion (as in: if you wanted to respect the user's targeting no matter what, willing to block the combo for it), but it will be able to hang the combo if it isn't done; probably better handled with Configs regarding what targets to include/how strictly to check them, if you did want to skip the range check.