fix(UI): ~Conflicts UI, for non-Combo types
This fixes the Conflicts UI (mostly for non-combo types of conflicts) to display the correct conflict message.
This also adjusts the target conflict UI to more clearly separate lines of conflicts.
This also adjusts the UI of conflict messages to have both lines centered, instead of the block as a whole being centered (leaving the first line of all conflict types being un-centered).