feat(Conflicts): +`ConflictCheck` abstract, +`MOAction` checking
This adds the `ConflictCheck` abstract class for all plugin conflict checking classes to inherit, that reduces some of the duplicate code required between those classes and helps enforce standard names for public Properties and consistent logging.
This also fits the `BossModCheck` class to that new inheritance.
This also adds the initial `MOAction` conflict checking (waiting on an attribute for the CCP to list out which actions are retargeted, specifically).