feat(UI): ~Left Column default size (+Minimum)
This changes the default size of the plugin's Left Column (the column with the tabs and icon) to be the width of the image, plus some.
This also changes the column to now have a minimum size (of just under the default size). When the user reduces the size below the minimum, the column sets itself to be unresizable (to stop the drag event) and sets itself to the default column width.