feat(UI): +Collapse Button
This adds a button to `ConfigWindow`, seen from anywhere in the plugin, that allows the user to collapse the sidebar entirely.
Additionally, a setting was added to store this collapsed state across plugin loads.

The main reason for adding this was to still allow tiny (read: nonexistent, now) sidebars, despite adding a minimum width to the sidebar in a previous commit.