feat(UI): ~Collapse Button
This makes the button point left or right, depending on if it is collapsed currently, changes the hover text depending on the same, and scales up the button if it is collapsed to be more noticeable.
Additionally this also corrects the window size, so it doesn't make UI elements below and to the right of it unclickable.